Changed model lists, separated test files.
