+++ text_generation/causal_lm/cpp/continuous_batching/python/python.cpp
@@ -4,6 +4,7 @@
 
 #include "pybind11/pybind11.h"
 #include <pybind11/stl.h>
+
 #include "continuous_batching_pipeline.hpp"
 
 namespace py = pybind11;
